% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_join_fun.R
\name{ACARA_vars}
\alias{ACARA_vars}
\title{Join in ACARA variables}
\usage{
ACARA_vars(data)
}
\description{
\code{ACARA_vars} returns school sector, ICSEA, and gelocation, columns from the ACARA series dataset, matched by 'latest_secondary_school_code'.
Warning: 'ACARA_vars' should only be used in cases where the student was in secondary school between 2015 & 2017. A more accurate match file needs to be developed that specifies year.
}
