% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_join_fun.R
\name{BOA_var}
\alias{BOA_var}
\title{Join in HEIMS basis of admission variable}
\usage{
BOA_var(data)
}
\description{
\code{BOA_var} returns a column of simplified basis of admission categories, consistent with the HEIMS definition, and joined by 'basis_for_admission_code'.
}
