% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheedr_palette.R
\docType{data}
\name{cbPalette}
\alias{cbPalette}
\title{Colour-blind friendly palette.}
\format{An object of class \code{character} of length 8.}
\usage{
cbPalette
}
\description{
#' \code{cbPalette} returns a colour palette of colour-blind friendly shades.
A vector of RBG colour codes, commonly used in 'scale' ggplot2 arguments that require colour values.
}
\keyword{datasets}
