% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_clean_fun.R
\name{Indigenous_var}
\alias{Indigenous_var}
\title{These functions perform common CHEEDR cleaning jobs when working
with La Trobe University enrolment data}
\usage{
Indigenous_var(data)
}
\description{
These functions perform common CHEEDR cleaning jobs when working
with La Trobe University enrolment data
}
