% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheedr_palette.R
\docType{data}
\name{LTU.Palette}
\alias{LTU.Palette}
\title{Colour palettes for CHEEDR work.}
\format{An object of class \code{character} of length 9.}
\usage{
LTU.Palette
}
\description{
\code{LTU.Palette} returns the colour palette using La Trobe University colours, reds, greys, oranges, and blacks.
}
\details{
\code{cbPalette} returns a colour palette of colour-blind friendly shades.

This is vector of colour codes, and is commonly used in 'scale' ggplot arguments that require colour code values.
}
\keyword{datasets}
