% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheedr_palette.R
\docType{data}
\name{LTU_Palette}
\alias{LTU_Palette}
\title{La Trobe colour palette.}
\format{An object of class \code{character} of length 9.}
\usage{
LTU_Palette
}
\description{
\code{LTU_Palette} returns a colour palette using La Trobe University colours, reds, greys, oranges, and blacks.
A vector of RGB colour codes, commonly used in 'scale' ggplot2 arguments that require colour values.
}
\keyword{datasets}
