% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_dat_funs.R
\name{new_csv_dir}
\alias{new_csv_dir}
\title{Set up directories and read in data
Create directory of folders for saving the csv BO extracts into (change default date to SysDate)
Consider including retention and completion data}
\usage{
new_csv_dir(dir_path = "~/", postcode = T, parent = T, date_stamp = Sys.Date())
}
\description{
Set up directories and read in data
Create directory of folders for saving the csv BO extracts into (change default date to SysDate)
Consider including retention and completion data
}
