% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parent_attain_function.R
\name{Parent_attain}
\alias{Parent_attain}
\title{Tranform parental highest attainment data from the
La Trobe University student enrolment database into tidy data
This function reads in a .csv of La Trobe enrolment parent attainment data,
and converts it into a standardised format
NOTE: You will need to at least include the following variables in the LTU BO extract:
# - 'Student ID'
# - 'Parent Highest Educ Attainment Code'}
\usage{
Parent_attain(path)
}
\description{
Tranform parental highest attainment data from the
La Trobe University student enrolment database into tidy data
This function reads in a .csv of La Trobe enrolment parent attainment data,
and converts it into a standardised format
NOTE: You will need to at least include the following variables in the LTU BO extract:
# - 'Student ID'
# - 'Parent Highest Educ Attainment Code'
}
