% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_join_fun.R
\name{Postcode_vars}
\alias{Postcode_vars}
\title{These functions join in common external data
(ASGS,SEIFA,country,language,basis of admission)}
\usage{
Postcode_vars(data)
}
\description{
These functions join in common external data
(ASGS,SEIFA,country,language,basis of admission)
}
