% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common_join_fun.R
\name{Postcode_vars}
\alias{Postcode_vars}
\title{\code{Postcode_vars} returns new columns to the df joined by 'postcode', including ASGS and SEIFA variables for 2011 and 2016.}
\usage{
Postcode_vars(data)
}
\description{
\code{Postcode_vars} returns new columns to the df joined by 'postcode', including ASGS and SEIFA variables for 2011 and 2016.
}
